Modularized heuristics
