+++ src/heuristics/simple.py
@@ -1,0 +1,13 @@
+from src.domain.state import State
+from src.heuristics.heuristic import Heuristic
+
+
+class HeuristicSimple(Heuristic):
+    def __init__(self, initial_state: 'State'):
+        super().__init__(initial_state)
+
+    def f(self, state):
+        return state
+
+    def __repr__(self):
+        return "Simple heuristic"
